backend/pyproject.toml
[project]
name = "thevault-backend"
version = "0.1.0"
description = "FastAPI backend for TheVault secrets manager"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.99.0,<0.100.0",
  "pg8000>=1.31.0,<2.0.0",
  "pydantic>=1.10.0,<2.0.0",
  "pyaes>=1.6.1,<2.0.0",
  "PyJWT>=2.10.0,<3.0.0",
  "redis>=6.0.0,<7.0.0",
  "uvicorn>=0.24.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.4.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=80.0.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
